+++ src/com/cloudogu/sos/pipebuildlib/stages/DoguStageModule.groovy
@@ -1,0 +1,5 @@
+package com.cloudogu.sos.pipebuildlib.stages
+
+interface DoguStageModule {
+    void register(DoguPipe pipe, StageGroup group)
+}
